bot/requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
requests==2.32.0
beautifulsoup4==4.12.3
psycopg2-binary==2.9.9
tenacity
